refactor(server): split server init versus routes plus error taxonomy
