+++ backend/internal/server/anthropic_errors.go
@@ -5,46 +5,8 @@ import (
 	"math"
 	"net/http"
 	"strconv"
-	"strings"
 	"time"
 )
-
-// quotaSummary renders the live per-model session quota from a probe's
-// isAnthropicRequest reports whether the incoming request is destined for the
-// Anthropic Messages surface (/v1/messages) or carries Anthropic headers.
-func isAnthropicRequest(r *http.Request) bool {
-	if r == nil {
-		return false
-	}
-	if strings.HasPrefix(r.URL.Path, "/v1/messages") {
-		return true
-	}
-	if r.Header.Get("anthropic-version") != "" || r.Header.Get("anthropic-api-key") != "" {
-		return true
-	}
-	return false
-}
-
-// anthropicErrorType maps HTTP status code and internal error code to standard
-// Anthropic error types per reference/protocols/anthropic-sdk-typescript.
-func anthropicErrorType(status int, code string) string {
-	switch {
-	case status == http.StatusUnauthorized:
-		return "authentication_error"
-	case status == http.StatusForbidden:
-		return "permission_error"
-	case status == http.StatusNotFound:
-		return "not_found_error"
-	case status == http.StatusTooManyRequests:
-		return "rate_limit_error"
-	case status == http.StatusServiceUnavailable && (code == "waiting_room_queued" || code == "waiting_room_required" || code == "capacity_deferred"):
-		return "overloaded_error"
-	case status >= 500:
-		return "api_error"
-	default:
-		return "invalid_request_error"
-	}
-}
 
 // writeAnthropicError writes an Anthropic-formatted error response:
 // {"type": "error", "error": {"type": "...", "message": "...", "code": "..."}}
