+++ backend/internal/server/middleware_logs.go
@@ -6,7 +6,6 @@ import (
 	"crypto/rand"
 	"errors"
 	"fmt"
-	"io/fs"
 	"net"
 	"net/http"
 	"strings"
